# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Pago de Nómina",
        "description": "Documentación generada con Swagger",
        "contact": {
            "name": "Soporte",
            "email": "tere_esti@fpuna.edu.py"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/empleados": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Listar empleados",
                "operationId": "f642e4a52968639a7b2bedd9ab443f78",
                "responses": {
                    "200": {
                        "description": "Lista de empleados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crear empleado",
                "operationId": "755511e2d2ece5cb9879b31f37a560b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Empleado"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado"
                    }
                }
            }
        },
        "/api/v1/empleados/{id}": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtener un empleado",
                "operationId": "96d4b280765d76e25d94a9851364e615",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado encontrado"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar un empleado",
                "operationId": "d6ba9d323d496883ef5fc4ae924c0b2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Empleado"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Eliminar un empleado",
                "operationId": "fabd7bdc966718dd543e154984f5ebef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado eliminado"
                    },
                    "404": {
                        "description": "Empleado no encontrado"
                    }
                }
            }
        },
...
        "/api/v1/nomina": {
            "get": {
                "tags": [
                    "Nominas"
                ],
                "summary": "Listar todas las nóminas",
                "operationId": "1afd0b3158660cd130ca4f67fd38213a",
                "responses": {
                    "200": {
                        "description": "Lista de nóminas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Nominas"
                ],
                "summary": "Crear una nueva nómina",
                "operationId": "7ddbfa3284e5daa3e8d08a00a848a4d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nomina"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nómina creada"
                    }
                }
            }
        },
        "/api/v1/nomina/{id}": {
            "get": {
                "tags": [
                    "Nominas"
                ],
                "summary": "Obtener una nómina específica",
                "operationId": "e14c0df544cf4181bb1dc02b707aa054",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nómina encontrada"
                    },
                    "404": {
                        "description": "Nómina no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Nominas"
                ],
                "summary": "Actualizar una nómina",
                "operationId": "ebac59e5a5da135ff1cbc94d0f2b6023",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Nomina"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nómina actualizada"
                    },
                    "404": {
                        "description": "Nómina no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Nominas"
                ],
                "summary": "Eliminar una nómina",
                "operationId": "cffc683c56a4e20eb1a0c55306be049a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nómina eliminada"
                    },
                    "404": {
                        "description": "Nómina no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Empleado": {
                "title": "Empleado",
                "description": "Modelo del empleado",
                "required": [
                    "id_persona",
                    "id_cargo",
                    "id_departamento",
                    "fecha_ingreso",
                    "salario_base",
                    "estado_empleado"
                ],
                "properties": {
                    "id_empleados": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "id_persona": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_cargo": {
                        "type": "integer",
                        "example": 2
                    },
                    "id_departamento": {
                        "type": "integer",
                        "example": 3
                    },
                    "fecha_ingreso": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-05-10"
                    },
                    "salario_base": {
                        "type": "number",
                        "format": "float",
                        "example": 2500.75
                    },
                    "estado_empleado": {
                        "type": "string",
                        "example": "Activo"
                    },
                    "fecha_egreso": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Nomina": {
                "required": [
                    "periodo",
                    "fecha_proceso_liquidacion",
                    "estado_nomina"
                ],
                "properties": {
                    "periodo": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-04-01"
                    },
                    "fecha_proceso_liquidacion": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-04-20"
                    },
                    "estado_nomina": {
                        "type": "string",
                        "enum": [
                            "Modificable",
                            "Confirmada"
                        ],
                        "example": "Modificable"
                    }
                },
                "type": "object"
            }
        }
    },
...
}